Actions Edit and Delete

+++ UniversityApp.BL/DTOs/CourseDTO.cs
@@ -5,12 +5,12 @@ namespace UniversityApp.BL.DTOs
     public class CourseDTO
     {
         [JsonProperty("CourseID")]
-        public long CourseID { get; set; }
+        public int CourseID { get; set; }
 
         [JsonProperty("Title")]
         public string Title { get; set; }
 
         [JsonProperty("Credits")]
-        public long Credits { get; set; }
+        public int Credits { get; set; }
     }
 }
